% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{AIC.tramvs}
\alias{AIC.tramvs}
\title{AIC "tramvs"}
\usage{
\method{AIC}{tramvs}(object, ...)
}
\arguments{
\item{object}{object of class \code{"tramvs"}}

\item{...}{additional arguments to \code{AIC()}}
}
\value{
Numeric vector containing AIC of best model
}
\description{
AIC "tramvs"
}
