% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.abess_tram}
\alias{coef.abess_tram}
\title{Coef "abess_tram"}
\usage{
\method{coef}{abess_tram}(object, ...)
}
\arguments{
\item{object}{object of class \code{"tramvs"}}

\item{...}{additional arguments to \code{coef()}}
}
\description{
Coef "abess_tram"
}
