% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.mmltvs}
\alias{coef.mmltvs}
\title{Coef "mmltvs"}
\usage{
\method{coef}{mmltvs}(object, best_only = FALSE, ...)
}
\arguments{
\item{object}{Object of class \code{"tramvs"}}

\item{best_only}{Wether to return the coefficients of the best model only
(default: FALSE)}

\item{...}{additional arguments to \code{coef()}}
}
\value{
Vector (\code{best_only = TRUE}) or matrix (\code{best_only = FALSE})
    of coefficients
}
\description{
Coef "mmltvs"
}
