% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tramvs.R
\name{cor_init.default}
\alias{cor_init.default}
\title{Default method for computing correlation}
\usage{
\method{cor_init}{default}(m0, mb)
}
\arguments{
\item{m0}{\code{modFUN(formula, data)}}

\item{mb}{\code{modFUN(mandatory, data)}}
}
\value{
Vector of correlation for initializing the active set
}
\description{
Default method for computing correlation
}
