% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmltvs.R
\name{cor_init.mmlt}
\alias{cor_init.mmlt}
\title{Method for computing correlations in mmlts}
\usage{
\method{cor_init}{mmlt}(m0, mb)
}
\arguments{
\item{m0}{\code{modFUN(formula, data)}}

\item{mb}{\code{modFUN(mandatory, data)}}
}
\value{
Vector of correlation for initializing the active set
}
\description{
Method for computing correlations in mmlts
}
