% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tramvs.R
\name{cor_init.stram}
\alias{cor_init.stram}
\title{Shit-scale tram method for computing correlation}
\usage{
\method{cor_init}{stram}(m0, mb)
}
\arguments{
\item{m0}{\code{modFUN(formula, data)}}

\item{mb}{\code{modFUN(mandatory, data)}}
}
\value{
Vector of correlations for initializing the active set, includes both
    shift and scale residuals
}
\description{
Shit-scale tram method for computing correlation
}
