% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.tramvs}
\alias{plot.tramvs}
\title{Plot "tramvs" object}
\usage{
\method{plot}{tramvs}(x, which = c("tune", "path"), ...)
}
\arguments{
\item{x}{object of class \code{"tramvs"}}

\item{which}{plotting either the regularization path (\code{"path"})
or the information criterion against the support size (\code{"tune"}, default)}

\item{...}{additional arguments to \code{plot()}}
}
\value{
Returns \code{invisible(NULL)}
}
\description{
Plot "tramvs" object
}
