% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.tramvs}
\alias{predict.tramvs}
\title{Predict "tramvs"}
\usage{
\method{predict}{tramvs}(object, ...)
}
\arguments{
\item{object}{object of class \code{"tramvs"}}

\item{...}{additional arguments to \code{predict.tram()}}
}
\value{
See \code{\link[tram]{predict.tram}}
}
\description{
Predict "tramvs"
}
