% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.tramvs}
\alias{print.tramvs}
\title{Print "tramvs"}
\usage{
\method{print}{tramvs}(x, ...)
}
\arguments{
\item{x}{object of class \code{"tramvs"}}

\item{...}{ignored}
}
\description{
Print "tramvs"
}
