% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{residuals.tramvs}
\alias{residuals.tramvs}
\title{Residuals "tramvs"}
\usage{
\method{residuals}{tramvs}(object, ...)
}
\arguments{
\item{object}{object of class \code{"tramvs"}}

\item{...}{additional arguments to \code{residuals()}}
}
\value{
Numeric vector containing residuals of best model,
     see \code{\link[tram]{residuals.tram}}
}
\description{
Residuals "tramvs"
}
