% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{SIC}
\alias{SIC}
\title{SIC generic}
\usage{
SIC(object, ...)
}
\arguments{
\item{object}{Model to compute SIC from}

\item{...}{for methods compatibility only}
}
\value{
Numeric vector (\code{best_only = TRUE}) or data.frame with SIC values
}
\description{
SIC generic
}
