% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{SIC.tramvs}
\alias{SIC.tramvs}
\title{SIC "tramvs"}
\usage{
\method{SIC}{tramvs}(object, best_only = FALSE, ...)
}
\arguments{
\item{object}{object of class \code{"tramvs"}}

\item{best_only}{Wether to return the coefficients of the best model only
(default: FALSE)}

\item{...}{for methods compatibility only}
}
\description{
SIC "tramvs"
}
