% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{simulate.tramvs}
\alias{simulate.tramvs}
\title{Simulate "tramvs"}
\usage{
\method{simulate}{tramvs}(object, nsim = 1, seed = NULL, ...)
}
\arguments{
\item{object}{object of class \code{"tramvs"}}

\item{nsim}{number of simulations}

\item{seed}{random seed for simulation}

\item{...}{additional arguments to \code{simulate()}}
}
\description{
Simulate "tramvs"
}
