% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.tramvs}
\alias{summary.tramvs}
\title{Summary "tramvs"}
\usage{
\method{summary}{tramvs}(object, ...)
}
\arguments{
\item{object}{object of class \code{"tramvs"}}

\item{...}{ignored}
}
\value{
\code{"tramvs"} object is returned invisibly
}
\description{
Summary "tramvs"
}
