% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{support.tramvs}
\alias{support.tramvs}
\title{Support "tramvs"}
\usage{
\method{support}{tramvs}(object, ...)
}
\arguments{
\item{object}{object of class \code{"tramvs"}}

\item{...}{ignored}
}
\value{
Character vector containing active set of best fit
}
\description{
Support "tramvs"
}
