% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{logLik.tramvs}
\alias{logLik.tramvs}
\title{logLik "tramvs"}
\usage{
\method{logLik}{tramvs}(object, ...)
}
\arguments{
\item{object}{object of class \code{"tramvs"}}

\item{...}{additional arguments to \code{logLik()}}
}
\description{
logLik "tramvs"
}
